refactor: prisma call for reports into server action
